+++ src/Extended.Dapper.Core/Sql/Generator/SqlGenerator.Insert.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using System.Reflection;
 using Extended.Dapper.Core.Attributes.Entities;
 using Extended.Dapper.Core.Attributes.Entities.Relations;
+using Extended.Dapper.Core.Database;
 using Extended.Dapper.Core.Mappers;
 using Extended.Dapper.Core.Sql.Query;
 using Extended.Dapper.Core.Sql.Query.Models;
@@ -45,6 +46,11 @@ namespace Extended.Dapper.Core.Sql.Generator
                     insertQuery.IdAlreadyPresent = true;
                 else if (autoValueType == typeof(Guid))
                     autoValueProperty.PropertyInfo.SetValue(entity, Guid.NewGuid());
+                else if (autoValueType == typeof(int))
+                {
+                    insertQuery.AutoIncrementKey = true;
+                    insertQuery.AutoIncrementField = autoValueProperty;
+                }
                 else
                     throw new NotImplementedException($"AutoValue for type {autoValueProperty.PropertyInfo.PropertyType} is not supported");
             }
@@ -52,7 +58,9 @@ namespace Extended.Dapper.Core.Sql.Generator
             insertQuery.Insert
                 .AddRange(entityMap.MappedPropertiesMetadata
                     .Where(x => !x.PropertyInfo.GetCustomAttributes<RelationAttributeBase>().Any()
-                        && x.PropertyInfo.GetCustomAttribute<IgnoreOnInsertAttribute>() == null)
+                        && x.PropertyInfo.GetCustomAttribute<IgnoreOnInsertAttribute>() == null
+                        && ( insertQuery.AutoIncrementField == null 
+                            || (insertQuery.AutoIncrementField != null && x.ColumnName != insertQuery.AutoIncrementField.ColumnName)))
                     .Select(p => {
                         insertQuery.Params.Add("p_" + p.ColumnName, p.PropertyInfo.GetValue(entity));