+++ src/Extended.Dapper.Core/Sql/QueryProviders/ISqlQueryProvider.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Data;
 using System.Linq.Expressions;
+using Extended.Dapper.Core.Database;
 using Extended.Dapper.Core.Mappers;
 using Extended.Dapper.Core.Sql.Generator;
 using Extended.Dapper.Core.Sql.Query;
@@ -9,6 +10,11 @@ namespace Extended.Dapper.Core.Sql.QueryProviders
 {
     public interface ISqlQueryProvider
     {
+        /// <summary>
+        /// The type of the database
+        /// </summary>
+        DatabaseProvider ProviderType { get; set; }
+
         /// <summary>
         /// Escapes a table name in the correct format
         /// </summary>